+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * A fix for forwarding metrics with gRPC using the kubernetes discoverer. Thanks, [androohan](https://github.com/androohan)!
 * Regenerate testing certs/CA that have expired and have broken tests. Thanks, [randallm](https://github.com/randallm)
 * The config field `trace_lightstep_access_token` is redacted if printed. Thanks [arnavdugar](https://github.com/arnavdugar)!
+* A fix for the -a flag in veneur-prometheus. Thanks, [dmarriner](https://github.com/dmarriner)
 
 # 14.1.0, 2021-03-16